+++ src/widgets/index.ts
@@ -2,4 +2,4 @@ export * from './JudgeOutputArea';
 export * from './JudgePanel';
 export * from './JudgeSubmissionArea';
 export * from './JudgeTerminal';
-export * from './JudgePreviewPanel';
+export * from './JudgeProblemPanel';
